Function for adding baseline matrices in SGPmatrices to SGPstateData in SGP package.

addBaselineMatrices(
    state,
    year,
    add.matrices.to.which.state=NULL)

Arguments

state
A character string for the state/assessment abbreviation for the baseline matrices to be used. Must match value to be used in SGPstateData.

year
Year associated with the matrices.

add.matrices.to.which.state
A character string for the state/assessment abbreviation where the baseline matrices will be placed in SGPstateData. If not argument is supplied the default is to use the state/assessment abbreviation supplied via the argument state.

Value

Returns SGPstateData with the matrices associated with argument state embedded in the position associated with the argument add.matrices.to.state (defaults to state).